Multiple changes to fix issue #18
- Shrunk `Session`, now is contained with in app.rs (may change this again in the future).
- `load_file_dialog` now uses `load_from_file` to reduce code duplication.

So what was causing the issue described in #18? Well, it turned out that when loading the file, it was a mistake to create a new `Arc`, instead the proper way to do load the new received data was to take the already pre-existing `Arc`, get the `RwLock` inside it and write to the `RwLock` rather than outright replacing the `Arc` itself.

For example:

```
// What we were doing before (approx.)
self.data.received = Arc::new(file_data.received) // all other references to the original `Arc` are now nothing because the original no longer exists, which caused the "not updating issue".

// What we're doing now
*self.data.received.write().unwrap() = file_data.received; // Original `Arc` is still in use, previous references are still valid and are now update properly.
```