Improve behaviour of username removal
Previously, the server only accounted for patterns based on the type of
operating system the _server_ was ran on, which is incredibly stupid.
Now, it checks all the patterns, sorted in most likely to least likely
to occur.

The regex patterns are lazy loaded, they are initialised for the first
time when `strip_username_from_path` is called for the first time and
will be re-used for the life of the server onwards.