Fix Windows username recognition
This change adds an additional pattern to censor Windows user directory strings with double backslashes, as not all loggers may strip these.

Also fixes an issue where Windows users with a space in their username would not be censored. For example, my username: "Samuel Boyden" would be "censored" as "<username> Boyden". This change fixes that by accounting for a space character in the Windows-based patterns.